refactor: 갈틱폰 폴더 구조 변경 및 파일 이름 변경
