+++ packages/backend/src/domains/garticphone/entity/albumData.ts
@@ -1,0 +1,13 @@
+import { DataType } from "../../../types/gartic.type";
+
+export class AlbumData {
+  type: DataType;
+  ownerId: string;
+  data: string;
+
+  constructor(type: DataType, ownerId: string, data: string) {
+    this.type = type;
+    this.ownerId = ownerId;
+    this.data = data;
+  }
+}
