+++ packages/backend/src/domains/garticphone/presenters/roomAPI.presenter.ts
@@ -1,7 +1,7 @@
-import { GarticphoneToRoom } from "../useCases/garticphoneToRoom.port";
+import { RoomAPIPort } from "../useCases/ports/roomAPI.port";
 import { DomainConnecter } from "../../../utils/domainConnecter";
 
-export class GarticphoneToRoomPresenter implements GarticphoneToRoom {
+export class RoomAPIPresenter implements RoomAPIPort {
   connecter = DomainConnecter.getInstance();
 
   gameEnded(roomId: string) {
